refactor: рефакторинг страницы 404 по замечаниям
